refactor(parser): split expr_pratt into submodules
The single 1200-line expr_pratt.rs mixed parser core, primary expression
parsing, and tests, making it hard to navigate.

The diagnostics module also duplicated its token-to-name mapping between
tok_name and a local expected_name helper.

Summary of changes:
- Split expr_pratt.rs into mod.rs, primary.rs, and tests.rs via include!
- Make diagnostics::expected_name the single canonical token-name match
  and derive tok_name from it instead of duplicating the match
- Remove PlusPlus/MinusMinus from prefix binding power (not prefix ops)
- Drop now-unneeded #[allow(dead_code)] attributes on used items
